+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * `distinct` summarizer as beed added. Collects distinct values as
   and array.
 
+* `percentile(:attr, nth)` summarizer as beed added. Collects the nth
+  percentile via a sort method (O(n) memory requirement!)
+
 * `Summarizer.by_proc(least){|t,memo| ... }` can now be used to factor a
   summarizer that works like `each_with_object`. `least` is the initial
   value, and defaults to nil.